rename entrypoint
entrypoint -> main
