+++ main.py
@@ -1,117 +1,172 @@
 #!/usr/bin/env python3
 
-import flet as ft
-from flet import (
-    Column,
-    Container,
-    Dropdown,
-    ElevatedButton,
-    RoundedRectangleBorder,
-    Page,
-    Row,
-    Text,
-    UserControl,
-    border_radius,
-    colors,
-    dropdown
+from nicegui import app, ui, Client
+from util import (
+    add_restaurant,
+    delete_restaurant,
+    get_all_restaurants,
+    get_restaurants,
+    rng_restaurant
 )
-import logging
-import random
-# import sqlite3
-# from datetime import datetime
-from util import create_db_and_tables, get_all_restaurants, get_restaurants, rng_restaurant
-from dataclasses import dataclass, field
-from decouple import config
 from pathlib import Path
-# from sqlmodel import Field, Session, SQLModel, create_engine, select
-from typing import Optional, Dict, List, Tuple
 
-# log level
-logging.basicConfig(level=logging.INFO)
+db_fn = Path(__file__).parent / "lunch.db"
 
 
-@dataclass
-class CustomButton():
-    text: str
-    style: Tuple = field(default_factory=lambda: ft.ButtonStyle(shape={
-        ft.MaterialState.DEFAULT: RoundedRectangleBorder(radius=2),
-    }))
-    on_click: Optional[ft.ElevatedButton().on_click] = None
+def random_restaurant():
+    """Return random restaurant based on cost."""
 
-    def __call__(self, text):
-        return ft.ElevatedButton(
-                text=self.text,
-                style=self.style,
-                on_click=self.on_click,
+    option = radio.value.lower()
+    restaurant = rng_restaurant(option)
+    ui.notify(
+        f"Today's {option} lunch is at {restaurant}",
+        position="center",
+        multi_line=True,
+        close_button=True,
+    )
+    # toggle input form visibility if open
+    add_input.set_visibility(False)
+    del_input.set_visibility(False)
+
+
+def all_restaurants():
+    """Return list of all restaurants."""
+
+    option = radio.value.lower()
+    restaurants = "\n".join([i for i in get_restaurants(option)])
+    ui.notify(
+        f"All {option} restaurants: {restaurants}",
+        position="center",
+        multi_line=True,
+        close_button=True,
+    )
+    # toggle input form visibility if open
+    add_input.set_visibility(False)
+    del_input.set_visibility(False)
+
+
+def update_database(*args):
+    """Add/remove restaurant from database."""
+
+    name = args[1]
+    option = radio.value.lower()
+
+    if args[0] == "add":
+        result = add_restaurant(name, option)
+        if result is False:
+            return ui.notify(
+                f"{name} already exists in database.",
+                position="center",
+                multi_line=True,
+                close_button=True,
+            )
+    elif args[0] == "remove":
+        result = delete_restaurant(name)
+        if result is None:
+            return ui.notify(
+                f"{name} does not exist in database.",
+                position="center",
+                multi_line=True,
+                close_button=True,
             )
 
+    if args[0] == "add":
+        msg = f"{name} has been {args[0]}ed."
+    elif args[0] == "remove":
+        msg = f"{name} has been {args[0]}d."
 
-class Lunch(UserControl):
-    def __init__(self, page: Page):
-        super().__init__()
-        self.page = page
-        self.choice = None
-        self.snack_bar = None
-        self._build()
-
-
-        if button.data == "roll":
-        def on_click(self, button):
-            self.roll_lunch(self.choice.value)
-        elif button.data == "delete":
-            self.delete_restaurant()
-        elif button.data == "add":
-            self.add_restaurant()
-        elif button.data == "list":
-            self.list_all()
-        self.snack_bar = ft.SnackBar(
-            content=ft.Text(f"{button.text} clicked!"),
-            duration=ft.Duration(milliseconds=5000),
-        )
-        page.add(self.snack_bar)
-        self.snack_bar.show()
+    ui.notify(
+        msg,
+        position="center",
+        multi_line=True,
+        close_button=True,
+    )
 
 
-    def _build(self):
-        page = self.page
-        page.title = "Lunch"
-        page.background_color = colors.WHITE
-        page.padding = 10
-
-        choice = ft.RadioGroup(
-            content=ft.Row([
-                ft.Container(ft.Radio(value="cheap", label="Cheap"), alignment=ft.alignment.center),
-                ft.Container(ft.Radio(value="normal", label="Normal"), alignment=ft.alignment.center),
-            ],
-            alignment=ft.MainAxisAlignment.CENTER,)
-        )
-        page.add(ft.Text("Click below to find out what's for Lunch:"), choice)
-
-        # TODO: use inline row declaration instead of custom button class (QA)
-        row = Row(
-            controls=[
-                Container(CustomButton(text="Roll Lunch")(page), alignment=ft.alignment.center, on_click=self.on_click, data="roll"),
-                Container(CustomButton(text="Delete Restaurant")(page), alignment=ft.alignment.center, on_click=self.on_click, data="delete"),
-                Container(CustomButton(text="Add Restaurant")(page), alignment=ft.alignment.center, on_click=self.on_click, data="add"),
-                Container(CustomButton(text="List All")(page), alignment=ft.alignment.center, on_click=self.on_click, data="list"),
-            ],
-            alignment=ft.MainAxisAlignment.CENTER,
-        )
-        page.add(row)
+# window
+app.native.window_args = {
+    'title': 'Lunch',
+    'resizable': True,
+}
+app.native.start_args['debug'] = True
 
 
-def main(page: Page):
-    page.title = "Lunch"
-    page.window_width = 650
-    page.window_height = 275
-    page.vertical_alignment = "center"
-    page.horizontal_alignment = "center"
+@ui.refreshable
+def get_radio_value():
+    """Return radio value (cheap/normal)."""
 
-    # create application instance
-    app = Lunch(page)
-
-    # add application's root control to page
-    page.add(app)
+    radio_props = "flex flex-row w-full justify-center items-center content-center"
+    global radio
+    radio = ui.radio(
+        ["Cheap", "Normal"],
+        value="Cheap").bind_enabled(lambda value: value).classes(radio_props)
+    return radio.value
 
 
-ft.app(target=main)
+@ui.page('/')
+async def index(client: Client):
+    """Main page."""
+
+    # tailwindcss / quasar classes
+    md_props = "w-full text-lg text-center"
+    ui.markdown("Click below to find out what's for **Lunch**").classes(md_props)
+
+    # radio selection
+    radio = get_radio_value()
+
+    # row of buttons
+    button_props = "flex flex-row w-full justify-center items-center content-center"
+    with ui.row().classes(button_props) as row:
+        with row:
+            # * display random restaurant on click
+            ui.button("Roll Lunch", on_click=random_restaurant)
+
+            # * add restaurant to database
+            add_btn = ui.button(
+                "Add Restaurant",
+                on_click=lambda: (add_input.set_visibility(True), del_input.set_visibility(False))
+            )
+
+            # * delete restaurant from database
+            del_btn = ui.button(
+                "Delete Restaurant",
+                on_click=lambda: (del_input.set_visibility(True), add_input.set_visibility(False))
+            )
+
+            # * list all restaurants
+            ui.button("List All", on_click=all_restaurants)
+
+    # TODO: clear input form on submit
+    with ui.row().classes(button_props) as form:
+        # * add restaurant
+        with form:
+            global add_input
+            add_input = ui.input(
+                label="Add Restaurant",
+                placeholder="Restaurant Name",
+                validation={'Input too long': lambda value: len(value) < 25}
+            ).props('clearable').on(
+                'keydown.enter',
+                lambda: (update_database("add", (add_input.value)), add_input.set_visibility(False))
+            )
+            add_input.set_visibility(False)
+
+        # * delete restaurant
+        with form:
+            global del_input
+            del_input = ui.input(
+                label="Remove Restaurant",
+                placeholder="Restaurant Name",
+                validation={'Input too long': lambda value: len(value) < 25}
+            ).props('clearable').on(
+                'keydown.enter',
+                lambda: (update_database("remove", (del_input.value)), del_input.set_visibility(False))
+            )
+            del_input.set_visibility(False)
+
+
+# website
+ui.run()
+
+# desktop
+# ui.run(native=True, window_size=(640, 300), fullscreen=False)
